requirements.txt
pynput>=1.7.6
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.10.0
faster-whisper>=0.10.0
pystray>=0.19.4
Pillow>=9.5.0
pyperclip>=1.8.2
